+++ gradle.properties
@@ -1,7 +1,7 @@
 # IntelliJ Platform Artifacts Repositories
 # -> https://plugins.jetbrains.com/docs/intellij/intellij-artifacts.html
 
-pluginGroup = com.github.mitchellolsthoorn.testgenie
+pluginGroup = nl.tudelft.ewi.se.ciselab.testgenie
 pluginName = TestGenie
 # SemVer format -> https://semver.org
 pluginVersion = 0.0.1
