--- src/main/kotlin/com/github/mitchellolsthoorn/testgenie/toolwindow/TestGenieToolWindowFactory.kt
@@ -1,40 +1,0 @@
-package com.github.mitchellolsthoorn.testgenie.toolwindow
-
-import com.github.mitchellolsthoorn.testgenie.services.CoverageToolWindowDisplayService
-import com.github.mitchellolsthoorn.testgenie.services.TestCaseDisplayService
-import com.intellij.openapi.components.service
-import com.intellij.openapi.project.Project
-import com.intellij.openapi.wm.ToolWindow
-import com.intellij.openapi.wm.ToolWindowFactory
-import com.intellij.ui.content.Content
-import com.intellij.ui.content.ContentFactory
-
-/**
- * This class is responsible for creating the UI of the TestGenie tool window.
- */
-class TestGenieToolWindowFactory : ToolWindowFactory {
-    /**
-     * Initialises the UI of the tool window.
-     */
-    override fun createToolWindowContent(project: Project, toolWindow: ToolWindow) {
-        val testGeniePanelWrapper = TestGenieToolWindow()
-        val contentFactory: ContentFactory = ContentFactory.SERVICE.getInstance()
-
-        val testCaseDisplayService = project.service<TestCaseDisplayService>()
-        toolWindow.contentManager.addContent(
-            contentFactory.createContent(
-                testCaseDisplayService.mainPanel, "Generated Tests", true
-            )
-        )
-
-        val content: Content = contentFactory.createContent(testGeniePanelWrapper.getContent(), "Parameters", false)
-
-        val visualisationService = project.service<CoverageToolWindowDisplayService>()
-        toolWindow.contentManager.addContent(content)
-        toolWindow.contentManager.addContent(
-            contentFactory.createContent(
-                visualisationService.mainPanel, "Coverage Visualisation", true
-            )
-        )
-    }
-}
